backend/requirements.txt
# Core Django Dependencies - Simplified for Testing
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.10

# Utilities
python-decouple==3.8
Pillow==10.4.0